actualizacionv1.2

+++ app/src/main/java/com/example/ejm1/Clases/Categoria.kt
@@ -2,7 +2,7 @@ package com.example.ejm1.Clases
 
 data class Categoria(
     val descripcion: String,
-    val idCategoria: Int,
+    val idCategoria: Long,
     val nombre: String,
     val imagen:String
